fix: migrate provider settings (#5205)
* chore: fix migrate api key

* chore: add logs

* chore: deprecate migration step

* chore: clean up

* fix: key apply
