+++ web-app/src/services/providers.ts
@@ -1,18 +1,51 @@
 import { models as providerModels } from 'token.js'
 import { mockModelProvider } from '@/mock/data'
-import { EngineManager, SettingComponentProps } from '@janhq/core'
+import {
+  EngineManagementExtension,
+  EngineManager,
+  ExtensionTypeEnum,
+  SettingComponentProps,
+} from '@janhq/core'
 import { ModelCapabilities } from '@/types/models'
 import { modelSettings } from '@/lib/predefined'
 import { fetchModels } from './models'
 import { ExtensionManager } from '@/lib/extension'
 
 export const getProviders = async (): Promise<ModelProvider[]> => {
+  const engines = !localStorage.getItem('migration_completed')
+    ? await ExtensionManager.getInstance()
+        .get<EngineManagementExtension>(ExtensionTypeEnum.Engine)
+        ?.getEngines()
+    : {}
   const builtinProviders = mockModelProvider.map((provider) => {
     let models = provider.models as Model[]
     if (Object.keys(providerModels).includes(provider.provider)) {
       const builtInModels = providerModels[
         provider.provider as unknown as keyof typeof providerModels
       ].models as unknown as string[]
+
+      if (engines && Object.keys(engines).length > 0) {
+        for (const [key, value] of Object.entries(engines)) {
+          const providerName = key.replace('google_gemini', 'gemini')
+          if (provider.provider !== providerName) continue
+          const engine = value[0] as
+            | {
+                api_key?: string
+                url?: string
+                engine?: string
+              }
+            | undefined
+          if (engine && 'api_key' in engine) {
+            const settings = provider?.settings.map((e) => {
+              if (e.key === 'api-key')
+                e.controller_props.value = (engine.api_key as string) ?? ''
+              return e
+            })
+
+            provider.settings = settings
+          }
+        }
+      }
 
       if (Array.isArray(builtInModels))
         models = builtInModels.map((model) => {
@@ -33,11 +66,15 @@ export const getProviders = async (): Promise<ModelProvider[]> => {
           } as Model
         })
     }
+
     return {
       ...provider,
       models,
     }
   })
+  if (engines && Object.keys(engines).length > 0) {
+    localStorage.setItem('migration_completed', 'true')
+  }
 
   const runtimeProviders: ModelProvider[] = []
